refactor(writers): split the shared helper module into cohesive submodules
